use long instead of uint

+++ sound_law/rl/Action.cpp
@@ -1,6 +1,6 @@
 #include <Action.h>
 
-Action::Action(uint action_id, uint before_id, uint after_id)
+Action::Action(long action_id, long before_id, long after_id)
 {
     this->action_id = action_id;
     this->before_id = before_id;
